# openapi/opencode.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "opencode",
    "description": "opencode api",
    "version": "0.0.3"
  },
  "paths": {
...
    "/question": {
      "get": {
        "operationId": "question.list",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "List pending questions",
        "description": "Get all pending question requests across all sessions.",
        "responses": {
          "200": {
            "description": "List of pending questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/QuestionRequest" }
                }
              }
            }
          }
        }
      }
    },
    "/question/{requestID}/reply": {
      "post": {
        "operationId": "question.reply",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "workspace",
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "requestID",
            "schema": { "type": "string", "pattern": "^que.*" },
            "required": true
          }
        ],
        "summary": "Reply to question request",
        "description": "Provide answers to a question request from the AI assistant.",
        "responses": {
          "200": {
            "description": "Question answered successfully",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotFoundError" }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "answers": {
                    "description": "User answers in order of questions (each answer is an array of selected labels)",
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/QuestionAnswer" }
                  }
                },
                "required": ["answers"]
              }
            }
          }
        }
      }
    },
    "/question/{requestID}/reject": {
      "post": {
        "operationId": "question.reject",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "workspace",
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "requestID",
            "schema": { "type": "string", "pattern": "^que.*" },
            "required": true
          }
        ],
        "summary": "Reject question request",
        "description": "Reject a question request from the AI assistant.",
        "responses": {
          "200": {
            "description": "Question rejected successfully",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotFoundError" }
              }
            }
          }
        }
      }
    },
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          }
        },
        "requestBody": { "content": { "application/json": { "schema": {} } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QuestionOption": {
        "type": "object",
        "properties": {
          "label": {
            "description": "Display text (1-5 words, concise)",
            "type": "string"
          },
          "description": {
            "description": "Explanation of choice",
            "type": "string"
          }
        },
        "required": ["label", "description"]
      },
      "QuestionInfo": {
        "type": "object",
        "properties": {
          "question": { "description": "Complete question", "type": "string" },
          "header": {
            "description": "Very short label (max 30 chars)",
            "type": "string"
          },
          "options": {
            "description": "Available choices",
            "type": "array",
            "items": { "$ref": "#/components/schemas/QuestionOption" }
          },
          "multiple": {
            "description": "Allow selecting multiple choices",
            "type": "boolean"
          },
          "custom": {
            "description": "Allow typing a custom answer (default: true)",
            "type": "boolean"
          }
        },
        "required": ["question", "header", "options"]
      },
      "QuestionRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "pattern": "^que.*" },
          "sessionID": { "type": "string", "pattern": "^ses.*" },
          "questions": {
            "description": "Questions to ask",
            "type": "array",
            "items": { "$ref": "#/components/schemas/QuestionInfo" }
          },
          "tool": {
            "type": "object",
            "properties": {
              "messageID": { "type": "string", "pattern": "^msg.*" },
              "callID": { "type": "string" }
            },
            "required": ["messageID", "callID"]
          }
        },
        "required": ["id", "sessionID", "questions"]
      },
...
      "QuestionAnswer": { "type": "array", "items": { "type": "string" } },
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": { "type": "string" },
              "additionalProperties": {}
            }
          },
          "success": { "type": "boolean", "const": false }
        },
        "required": ["data", "errors", "success"]
      },
...
      "NotFoundError": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "const": "NotFoundError" },
          "data": {
            "type": "object",
            "properties": { "message": { "type": "string" } },
            "required": ["message"]
          }
        },
        "required": ["name", "data"]
      },
...
    }
  }
}